added button 'add watcher' to JSON wacthers page, little text and styling changes in the PageBase component
