--- app/renderer/pages/Help/index.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import PageBase from '../PageBase';
-
-const Help = () => {
-  return (
-    <PageBase>I am help page</PageBase>
-  );
-};
-
-export default Help;
