+++ app/renderer/pages/HelpPage/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import PageBase from '../common/PageBase';
+
+const Help = () => {
+  return (
+    <PageBase>I am help page</PageBase>
+  );
+};
+
+export default Help;
