+++ app/renderer/pages/JsonPage/JsonPageComponent.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import PageBase from '../common/PageBase';
+import { Button } from '@material-ui/core';
+
+const JsonPage = ({ handleAddWatcher }) => {
+  return (
+    <PageBase>
+      <Button variant='contained' color='primary' onClick={handleAddWatcher}>
+        Add watcher
+      </Button>
+    </PageBase>
+  );
+};
+
+export default JsonPage;
