+++ app/renderer/pages/JsonPage/useStyles.js
@@ -1,0 +1,8 @@
+import { makeStyles } from "@material-ui/styles";
+
+const useStyles = makeStyles(theme => ({
+    button: {
+    }
+}));
+
+export default useStyles;
