--- app/renderer/pages/PageBase.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Container } from '@material-ui/core';
-
-const PageBase = (props) => {
-  return (
-    <Container maxWidth="lg">
-      {props.children}
-    </Container>
-  );
-};
-
-export default PageBase;
