--- app/renderer/pages/Pages.js
@@ -1,6 +1,0 @@
-const Pages = {
-  Help: '/',
-  WebApp: '/webapp'
-};
-
-export default Pages;
