--- app/renderer/pages/WebApp/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import PageBase from '../PageBase';
-
-const WebApp = () => {
-  return (
-    <PageBase>I am webapp page !</PageBase>
-  );
-};
-
-export default WebApp;
