+++ app/renderer/pages/common/PageBase/useStyles.js
@@ -1,0 +1,9 @@
+import { makeStyles } from "@material-ui/styles";
+
+const useStyles = makeStyles(theme => ({
+    container: {
+        margin: '1.25rem'
+    }
+}));
+
+export default useStyles;
