+++ app/renderer/pages/common/Pages.js
@@ -1,0 +1,6 @@
+const Pages = {
+  Help: '/help',
+  Json: '/'
+};
+
+export default Pages;
